#142 BindingAdapter 클래스 분리
