--- app/src/main/java/com/webrtc/boyj/extension/databinding/BindingAdapters.java
@@ -1,101 +1,0 @@
-package com.webrtc.boyj.extension.databinding;
-
-import android.annotation.SuppressLint;
-import android.content.res.Resources;
-import android.view.View;
-import android.view.animation.Animation;
-import android.view.animation.Transformation;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.constraintlayout.widget.ConstraintLayout;
-import androidx.databinding.BindingAdapter;
-import androidx.recyclerview.widget.ListAdapter;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.webrtc.boyj.api.boyjrtc.BoyjMediaStream;
-import com.webrtc.boyj.extension.custom.BoyjSurfaceView;
-import com.webrtc.boyj.extension.custom.SplitLayout;
-import com.webrtc.boyj.presentation.call.CallAdapter;
-
-import org.webrtc.MediaStream;
-import org.webrtc.VideoTrack;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class BindingAdapters {
-    @SuppressWarnings("unchecked")
-    @BindingAdapter({"items"})
-    public static <T, VH extends RecyclerView.ViewHolder> void setItems(
-            @NonNull final RecyclerView recyclerView,
-            @Nullable final List<T> items) {
-        final ListAdapter<T, VH> adapter = (ListAdapter<T, VH>) recyclerView.getAdapter();
-        if (adapter != null) {
-            adapter.submitList(items == null ? null : new ArrayList<>(items));
-        }
-    }
-
-    @SuppressLint("DefaultLocale")
-    @BindingAdapter({"callTime"})
-    public static void setCallTime(@NonNull final TextView textView, final int time) {
-        if (time > 0) {
-            int min = time / 60;
-            int sec = time % 60;
-            textView.setVisibility(View.VISIBLE);
-            textView.setText(String.format("%02d:%02d", min, sec));
-        }
-    }
-
-    @BindingAdapter({"localStream"})
-    public static void bindMediaStream(@NonNull final BoyjSurfaceView surfaceView,
-                                       @Nullable final MediaStream mediaStream) {
-        if (mediaStream != null) {
-            final VideoTrack track = mediaStream.videoTracks.get(0);
-            surfaceView.setZOrderMediaOverlay(true);
-            track.addSink(surfaceView);
-        }
-    }
-
-    @BindingAdapter({"remoteStream"})
-    public static void bindMediaStream(@NonNull final SplitLayout splitLayout,
-                                       @Nullable final BoyjMediaStream mediaStream) {
-        final CallAdapter adapter = (CallAdapter) splitLayout.getAdapter();
-        if (adapter != null && mediaStream != null) {
-            adapter.addMediaStream(mediaStream);
-        }
-    }
-
-    @BindingAdapter({"leaveStream"})
-    public static void unbindMediaStream(@NonNull final SplitLayout splitLayout,
-                                         @Nullable final String leavedUserName) {
-        final CallAdapter adapter = (CallAdapter) splitLayout.getAdapter();
-        if (adapter != null && leavedUserName != null) {
-            adapter.removeStreamById(leavedUserName);
-        }
-    }
-
-    @BindingAdapter({"callAnimation"})
-    public static void callAnimation(@NonNull final BoyjSurfaceView view,
-                                     final boolean isCalling) {
-        if (isCalling) {
-            view.startAnimation(new Animation() {
-                @Override
-                protected void applyTransformation(float interpolatedTime, Transformation t) {
-                    final ConstraintLayout.LayoutParams params = (ConstraintLayout.LayoutParams) view.getLayoutParams();
-                    params.width = dp2Px(80);
-                    params.height = dp2Px(120);
-                    params.endToEnd = ConstraintLayout.LayoutParams.UNSET;
-                    params.bottomToBottom = ConstraintLayout.LayoutParams.UNSET;
-                    params.setMargins(dp2Px(24), dp2Px(24), 0, 0);
-                    view.setLayoutParams(params);
-                }
-            });
-        }
-    }
-
-    private static int dp2Px(int dp) {
-        return (int) (dp * Resources.getSystem().getDisplayMetrics().density);
-    }
-}
