+++ app/src/main/java/com/webrtc/boyj/presentation/common/binding/ListBinding.java
@@ -1,0 +1,23 @@
+package com.webrtc.boyj.presentation.common.binding;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.databinding.BindingAdapter;
+import androidx.recyclerview.widget.ListAdapter;
+import androidx.recyclerview.widget.RecyclerView;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ListBinding {
+    @SuppressWarnings("unchecked")
+    @BindingAdapter({"items"})
+    public static <T, VH extends RecyclerView.ViewHolder> void setItems(
+            @NonNull final RecyclerView recyclerView,
+            @Nullable final List<T> items) {
+        final ListAdapter<T, VH> adapter = (ListAdapter<T, VH>) recyclerView.getAdapter();
+        if (adapter != null) {
+            adapter.submitList(items == null ? null : new ArrayList<>(items));
+        }
+    }
+}
